# Faster enum impl