… hikari/messages.py
… hikari/utilities/cache.py
… hikari/utilities/enums.py
… hikari/webhooks.py
… scripts/enum_benchmark.py
… tests/hikari/utilities/test_enums.py