- If a value is failed to be found, will just return the inputed value. This way, if Discord adds new fields, it wont break.

Co-authored-by: @nekokatt

### Checklist
<!-- Make sure to tick all the following boxes by putting an `x` in between (like this `[x]`) -->
- [x] I have run `nox` and all the pipelines have passed.
- [x] I have made unittests according to the code I have added/modified/deleted.

### Related issues
Closes #182
Closes #151

# Performance metrics
```
BasicPyEnum.__call__('25') 0.43657039100071415 µs
BasicHikariEnum.__call__('25') 0.19885178900040046 µs
BasicPyEnum._value2member_map_['25'] 0.09891081800014945 µs
BasicHikariEnum._value2member_map['25'] 0.1033127640002931 µs
BasicPyEnum.__getitem__['z'] 0.16736914299872296 µs
BasicHikariEnum.__getitem__['z'] 0.15685381999901438 µs
BasicPyEnum.__call__ profile
         2000003 function calls in 0.744 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.249    0.249    0.744    0.744 <string>:1(<module>)
  1000000    0.274    0.000    0.495    0.000 enum.py:283(__call__)
  1000000    0.221    0.000    0.221    0.000 enum.py:562(__new__)
        1    0.000    0.000    0.744    0.744 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}


BasicHikariEnum.__call__ profile
         1000003 function calls in 0.340 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.183    0.183    0.340    0.340 <string>:1(<module>)
  1000000    0.157    0.000    0.157    0.000 enums.py:115(__call__)
        1    0.000    0.000    0.340    0.340 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}


BasicPyEnum.__getitem__ profile
         1000003 function calls in 0.331 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.165    0.165    0.331    0.331 <string>:1(<module>)
  1000000    0.166    0.000    0.166    0.000 enum.py:348(__getitem__)
        1    0.000    0.000    0.331    0.331 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}


BasicHikariEnum.__getitem__ profile
         1000003 function calls in 0.318 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.159    0.159    0.318    0.318 <string>:1(<module>)
  1000000    0.158    0.000    0.158    0.000 enums.py:141(__getitem__)
        1    0.000    0.000    0.318    0.318 {built-in method builtins.exec}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```