@FasterSpeeding 2020-09-22
Do we have some performance metrics on how faster this is?