@davfsa 2020-09-22
> Do we have some performance metrics on how faster this is compared to the stdlib enum class?

Edited the top comment to include them