@nekokatt 2020-09-23
This kinda links in to what I said in #201 or one of the related issues about whether we are misusing enums for sone stuff.

Lgtm though.