@nekokatt 2020-09-23
Just fix da merge conflicts and we can see.

I'd be interested to see how this affects the startup overhead with asyncio_debug=1